requirements.txt
langchain==0.3.21
langchain_chroma==0.2.2
langchain_community==0.3.20
langchain_core==0.3.47
langchain_huggingface==0.1.2
langchain_openai==0.3.9
openai==1.65.0
pdfplumber==0.11.5
python-dotenv==1.0.1
streamlit==1.43.2
torch==2.6.0
shutil
